requirements.txt
flask==3.1.1
python-dotenv==1.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
MarkupSafe==3.0.2 
Flask-Mail 